Save renderer settings to file
