+++ .gitignore
@@ -6,3 +6,4 @@
 save.bin
 save.txt
 rg3d.log
+settings.json
