--- settings.json
@@ -1,1 +1,0 @@
-{"move_forward":{"description":"Move Forward","button":{"Key":"W"}},"move_backward":{"description":"Move Backward","button":{"Key":"S"}},"move_left":{"description":"Move Left","button":{"Key":"A"}},"move_right":{"description":"Move Right","button":{"Key":"D"}},"jump":{"description":"Jump","button":{"Key":"Space"}},"crouch":{"description":"Crouch","button":{"Key":"C"}},"ads":{"description":"Aim Down Sights","button":{"Mouse":3}},"shoot":{"description":"Shoot","button":{"Mouse":1}},"next_weapon":{"description":"Next Weapon","button":"WheelUp"},"prev_weapon":{"description":"Previous Weapon","button":"WheelDown"},"run":{"description":"Run","button":{"Key":"LShift"}},"mouse_sens":0.2,"mouse_y_inverse":false,"smooth_mouse":false,"shake_camera":true}
