+++ src/control_scheme.rs
@@ -1,5 +1,4 @@
 use rg3d::event::VirtualKeyCode;
-use rg3d::utils::log::Log;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -28,13 +27,13 @@ impl ControlButton {
     }
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ControlButtonDefinition {
     pub description: String,
     pub button: ControlButton,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ControlScheme {
     pub move_forward: ControlButtonDefinition,
     pub move_backward: ControlButtonDefinition,
@@ -109,33 +108,6 @@ impl Default for ControlScheme {
 }
 
 impl ControlScheme {
-    pub fn load_from_file(filename: &str) -> Self {
-        if let Ok(Ok(settings)) = std::fs::read_to_string(std::path::Path::new(filename))
-            .as_ref()
-            .and_then(|f| serde::export::Ok(serde_json::from_str(f)))
-        {
-            Log::writeln("Successfully loaded settings".to_string());
-            settings
-        } else {
-            // Unable to read settings file, so fall back to defaults
-            Log::writeln(format!(
-                "Could not read settings file {} (missing or corrupted?), falling back to defaults",
-                filename
-            ));
-            Self::default()
-        }
-    }
-
-    pub fn write_to_file(&self, filename: &str) {
-        if let Err(error) = serde_json::to_string(self).and_then(|data| {
-            serde::export::Ok(std::fs::write(std::path::Path::new(filename), data))
-        }) {
-            Log::writeln(format!("Error saving settings: {}", error))
-        } else {
-            Log::writeln(format!("Succesfully saved settings to {}", filename));
-        }
-    }
-
     pub fn buttons_mut(&mut self) -> [&mut ControlButtonDefinition; 11] {
         [
             &mut self.move_forward,
